# Optimize tag handling