--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 - Put unreleased items here.
 
+## [2.0.0] - 2020-05-28
+
+- Go module support
+- Go 1.12-1.14 support
+- Upgrade to datadog-go 3.x
+- Adds support for DataDog Distributions: https://docs.datadoghq.com/metrics/distributions/
+- Adds Close() to Client interface to support intentional flushes when applications are shutting down
+- Optimize tag handling
+
 ## [1.4.0] - 2019-08-26
 
 - Updated `datadog-go` to version `2.2.0`