--- a/go.mod
+++ b/go.mod
@@ -3,10 +3,11 @@ module github.com/istreamlabs/go-metrics
 go 1.12
 
 require (
-	github.com/DataDog/datadog-go v3.4.1+incompatible
+	github.com/DataDog/datadog-go v3.7.1+incompatible
 	github.com/mattn/go-colorable v0.1.6 // indirect
 	github.com/mattn/go-isatty v0.0.12
 	github.com/mgutz/ansi v0.0.0-20170206155736-9520e82c474b
 	github.com/stretchr/testify v1.5.1 // indirect
 	golang.org/x/sys v0.0.0-20200302150141-5c8b2ff67527 // indirect
+	gopkg.in/go-playground/assert.v1 v1.2.1
 )