… Gopkg.lock
… go.sum